Migrate to net8.0, Update packages,
Nothing much changed, removed obsolete elements and removed castle build.
