--- Help/SimpleFeedReaderHelp.shfbproj
@@ -1,87 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003" ToolsVersion="4.0">
-  <PropertyGroup>
-    <!-- The configuration and platform will be used to determine which
-         assemblies to include from solution and project documentation
-         sources -->
-    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
-    <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
-    <SchemaVersion>2.0</SchemaVersion>
-    <ProjectGuid>{d6a78785-3434-4544-a495-b201afbd0964}</ProjectGuid>
-    <SHFBSchemaVersion>2017.9.26.0</SHFBSchemaVersion>
-    <!-- AssemblyName, Name, and RootNamespace are not used by SHFB but Visual
-         Studio adds them anyway -->
-    <AssemblyName>Documentation</AssemblyName>
-    <RootNamespace>Documentation</RootNamespace>
-    <Name>Documentation</Name>
-    <!-- SHFB properties -->
-    <OutputPath>.\Output\</OutputPath>
-    <HtmlHelpName>Documentation</HtmlHelpName>
-    <Language>en-US</Language>
-    <CopyrightHref>http://robiii.nl</CopyrightHref>
-    <CopyrightText>%28C%29 2014-2018 Devcorner.nl</CopyrightText>
-    <FeedbackEMailAddress>rob%40devcorner.nl</FeedbackEMailAddress>
-    <FeedbackEMailLinkText>Rob Janssen</FeedbackEMailLinkText>
-    <HelpTitle>SimpleFeedReader documentation</HelpTitle>
-    <RootNamespaceContainer>True</RootNamespaceContainer>
-    <RootNamespaceTitle>SimpleFeedReader documentation</RootNamespaceTitle>
-    <SelfBranded>False</SelfBranded>
-    <Preliminary>False</Preliminary>
-    <DocumentationSources>
-      <DocumentationSource sourceFile="..\SimpleFeedReader\bin\Release\SimpleFeedReader.dll" />
-      <DocumentationSource sourceFile="..\SimpleFeedReader\bin\Release\SimpleFeedReader.xml" />
-    </DocumentationSources>
-    <ProjectSummary>&amp;lt%3bpara&amp;gt%3bThe SimpleFeedReader library is used for retrieving syndication feeds in a quick and easy manner abstracting details and normalization of items even further than the .Net&amp;#39%3bs System.Services.Syndication does.&amp;lt%3b/para&amp;gt%3b</ProjectSummary>
-    <HelpFileVersion>1.0.8.0</HelpFileVersion>
-    <MaximumGroupParts>2</MaximumGroupParts>
-    <NamespaceGrouping>False</NamespaceGrouping>
-    <SyntaxFilters>Standard</SyntaxFilters>
-    <SdkLinkTarget>Blank</SdkLinkTarget>
-    <PresentationStyle>VS2013</PresentationStyle>
-    <NamingMethod>Guid</NamingMethod>
-    <ContentPlacement>AboveNamespaces</ContentPlacement>
-    <NamespaceSummaries>
-      <NamespaceSummaryItem name="SimpleFeedReader" isDocumented="True" xmlns="">&lt;para&gt;Contains classes and interfaces for the SimpleFeedReader library.&lt;/para&gt;</NamespaceSummaryItem>
-    </NamespaceSummaries>
-    <PlugInNamespaces>ms.vsipcc+, ms.vsexpresscc+</PlugInNamespaces>
-    <CollectionTocStyle>Hierarchical</CollectionTocStyle>
-    <MSHelp2SdkLinkType>Msdn</MSHelp2SdkLinkType>
-    <IncludeStopWordList>True</IncludeStopWordList>
-    <HelpFileFormat>HtmlHelp1</HelpFileFormat>
-    <FrameworkVersion>Cross-platform (.NET Core/.NET Standard)</FrameworkVersion>
-    <CleanIntermediates>True</CleanIntermediates>
-    <KeepLogFile>True</KeepLogFile>
-    <DisableCodeBlockComponent>False</DisableCodeBlockComponent>
-    <IndentHtml>False</IndentHtml>
-    <BuildAssemblerVerbosity>OnlyWarningsAndErrors</BuildAssemblerVerbosity>
-    <SaveComponentCacheCapacity>100</SaveComponentCacheCapacity>
-  </PropertyGroup>
-  <!-- There are no properties for these groups.  AnyCPU needs to appear in
-       order for Visual Studio to perform the build.  The others are optional
-       common platform types that may appear. -->
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|x86' ">
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|x86' ">
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|x64' ">
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|x64' ">
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|Win32' ">
-  </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|Win32' ">
-  </PropertyGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\SimpleFeedReader\SimpleFeedReader.csproj">
-      <Name>SimpleFeedReader</Name>
-      <Project>{b057867d-0d85-4eec-87c2-fba8020ed31d}</Project>
-      <Private>True</Private>
-    </ProjectReference>
-  </ItemGroup>
-  <!-- Import the SHFB build targets -->
-  <Import Project="$(SHFBROOT)\SandcastleHelpFileBuilder.targets" />
-</Project>
