+++ N2.RssAtomFeedReader.UnitTests/TestFeeds/basic.atom
@@ -1,0 +1,29 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<feed xmlns="http://www.w3.org/2005/Atom">
+  <title>Example Feed</title>
+  <link href="http://example.org/"/>
+  <updated>2014-04-16T13:57:35Z</updated>
+  <author>
+    <name>John Doe</name>
+  </author>
+  <id>urn:uuid:abf81fc9-76db-4e46-a83d-e05ec7f45ccb</id>
+
+  <entry>
+    <title>Test1</title>
+    <link href="http://example.org/foo/bar/1"/>
+    <id>urn:uuid:0ea6c57b-4546-4264-8b96-13434c349d87</id>
+    <published>2013-03-13T13:37:31Z</published>
+    <updated>2014-04-16T13:57:35Z</updated>
+    <summary>Summary1</summary>
+    <content type="html">
+      &lt;p&gt;HTML content&lt;/p&gt;
+    </content>
+  </entry>
+  <entry>
+    <title>Test2</title>
+    <link href="http://example.org/foo/bar/2"/>
+    <id>urn:uuid:d58672c4-f62e-483e-ab00-ff0940113e29</id>
+    <summary>Summary2</summary>
+    <content type="text">Text content</content>
+  </entry>
+</feed>
