+++ N2.RssAtomFeedReader.UnitTests/TestFeeds/decoding.rss
@@ -1,0 +1,23 @@
+﻿<?xml version="1.0" encoding="UTF-8"?>
+<rss version="2.0">
+  <channel>
+    <item>
+      <title>foo&amp;amp;amp;amp;amp;amp;bar</title>
+      <link>foo&amp;amp;amp;amp;amp;amp;bar</link>
+      <description>
+        foo&amp;amp;amp;amp;amp;amp;bar
+      </description>
+    </item>
+    <item>
+      <title>foo&amp;amp;amp;amp;amp;bar</title>
+      <link>foo&amp;amp;amp;amp;amp;bar</link>
+      <description>
+        foo&amp;amp;amp;amp;amp;bar
+      </description>
+    </item>
+    <item>
+      <title>&#x1EBF;</title>
+      <description>&#x0065;&#x0302;&#x0301;</description>
+    </item>
+  </channel>
+</rss>
