+++ N2.RssAtomFeedReader/Builder.cs
@@ -1,0 +1,8 @@
+﻿using System.Reflection;
+
+[assembly: AssemblyVersion("8.0.9")]
+
+namespace N2.FeedReader;
+public class Builder
+{
+}
