+++ N2.RssAtomFeedReader/IFeedItemNormalizer.cs
@@ -1,0 +1,18 @@
+﻿using System.ServiceModel.Syndication;
+
+namespace N2.RssAtomFeedReader
+{
+    /// <summary>
+    /// Provides the base FeedItemNormalizer interface for <see cref="FeedItem"/>s.
+    /// </summary>
+    public interface IFeedItemNormalizer
+    {
+        /// <summary>
+        /// Normalizes a <see cref="SyndicationItem"/> into a <see cref="FeedItem"/>.
+        /// </summary>
+        /// <param name="feed">The <see cref="SyndicationFeed"/> on which the item was retrieved.</param>
+        /// <param name="item">The <see cref="SyndicationItem"/> to normalize.</param>
+        /// <returns>Returns a normalized <see cref="FeedItem"/>.</returns>
+        FeedItem Normalize(SyndicationFeed feed, SyndicationItem item);
+    }
+}
