+++ SimpleFeedReader.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30011.22
+# Visual Studio Version 17
+VisualStudioVersion = 17.12.35514.174
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{35CEBD40-E6BE-4671-B270-89C7E72CA980}"
 	ProjectSection(SolutionItems) = preProject
@@ -9,14 +9,9 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution 
 		README.md = README.md
 	EndProjectSection
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SimpleFeedReader", "SimpleFeedReader\SimpleFeedReader.csproj", "{B057867D-0D85-4EEC-87C2-FBA8020ED31D}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "N2.RssAtomFeedReader", "N2.RssAtomFeedReader\N2.RssAtomFeedReader.csproj", "{003A8F2F-F910-4FD4-AF49-C8C7D2042668}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "SimpleFeedReaderTests", "SimpleFeedReaderTests\SimpleFeedReaderTests.csproj", "{F89D25D8-DDE6-4533-85BE-EBF105269C31}"
-EndProject
-Project("{7CF6DF6D-3B04-46F8-A40B-537D21BCA0B4}") = "SimpleFeedReaderHelp", "Help\SimpleFeedReaderHelp.shfbproj", "{D6A78785-3434-4544-A495-B201AFBD0964}"
-	ProjectSection(ProjectDependencies) = postProject
-		{B057867D-0D85-4EEC-87C2-FBA8020ED31D} = {B057867D-0D85-4EEC-87C2-FBA8020ED31D}
-	EndProjectSection
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "N2.RssAtomFeedReader.UnitTests", "N2.RssAtomFeedReader.UnitTests\N2.RssAtomFeedReader.UnitTests.csproj", "{B1836759-F532-4A22-AA0A-D0CA3E2B2715}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -24,17 +19,14 @@ Global
 		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{B057867D-0D85-4EEC-87C2-FBA8020ED31D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B057867D-0D85-4EEC-87C2-FBA8020ED31D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B057867D-0D85-4EEC-87C2-FBA8020ED31D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B057867D-0D85-4EEC-87C2-FBA8020ED31D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F89D25D8-DDE6-4533-85BE-EBF105269C31}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F89D25D8-DDE6-4533-85BE-EBF105269C31}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F89D25D8-DDE6-4533-85BE-EBF105269C31}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F89D25D8-DDE6-4533-85BE-EBF105269C31}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D6A78785-3434-4544-A495-B201AFBD0964}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D6A78785-3434-4544-A495-B201AFBD0964}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D6A78785-3434-4544-A495-B201AFBD0964}.Release|Any CPU.Build.0 = Release|Any CPU
+		{003A8F2F-F910-4FD4-AF49-C8C7D2042668}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{003A8F2F-F910-4FD4-AF49-C8C7D2042668}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{003A8F2F-F910-4FD4-AF49-C8C7D2042668}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{003A8F2F-F910-4FD4-AF49-C8C7D2042668}.Release|Any CPU.Build.0 = Release|Any CPU
+		{B1836759-F532-4A22-AA0A-D0CA3E2B2715}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B1836759-F532-4A22-AA0A-D0CA3E2B2715}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B1836759-F532-4A22-AA0A-D0CA3E2B2715}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B1836759-F532-4A22-AA0A-D0CA3E2B2715}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
