--- SimpleFeedReader/SimpleFeedReader.csproj
@@ -1,41 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFrameworks>net4;netstandard20</TargetFrameworks>
-    <AssemblyVersion>1.0.9.0</AssemblyVersion>
-    <FileVersion>1.0.9.0</FileVersion>
-    <Version>1.0.9.0</Version>
-    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-    <Authors>RobIII</Authors>
-    <Description>Easy to use, simple, Syndication feed reader (Atom / RSS).</Description>
-    <Summary>Provides classes and methods for consuming syndication feeds.</Summary>
-    <PackageId>SimpleFeedReader</PackageId>
-    <AssemblyName>SimpleFeedReader</AssemblyName>
-    <RootNamespace>SimpleFeedReader</RootNamespace>
-    <PackageProjectUrl>https://github.com/RobThree/SimpleFeedReader</PackageProjectUrl>
-    <RepositoryUrl>https://github.com/RobThree/SimpleFeedReader</RepositoryUrl>
-    <PackageTags>rss atom syndication feed</PackageTags>
-    <PackageReleaseNotes>* Added netcore support</PackageReleaseNotes>
-    <GenerateAssemblyInfo>false</GenerateAssemblyInfo>
-    <PackageIcon>icon.png</PackageIcon>
-    <Copyright>(c) 2014 - 2020 Devcorner.nl</Copyright>
-    <PackageLicenseExpression>MIT</PackageLicenseExpression>
-    <AutoGenerateBindingRedirects>true</AutoGenerateBindingRedirects>
-  </PropertyGroup>
-
-  <PropertyGroup Condition="'$(Configuration)|$(TargetFramework)|$(Platform)'=='Release|net4|AnyCPU'">
-    <OutputPath></OutputPath>
-    <DocumentationFile>bin\Release\SimpleFeedReader.xml</DocumentationFile>
-  </PropertyGroup>
-
-
-  <ItemGroup>
-    <None Include="..\Gfx\icon.png" Pack="true" PackagePath="" />
-  </ItemGroup> 
-  
-  <ItemGroup>
-    <PackageReference Include="System.ServiceModel.Syndication" Version="4.7.0" Condition="'$(TargetFramework)' == 'netstandard20'" />
-    <Reference Include="System.ServiceModel" Condition="'$(TargetFramework)' == 'net4'" />
-  </ItemGroup> 
-
-</Project>
