--- SimpleFeedReaderTests/TestFeeds/xml_injection1.rss
@@ -1,15 +1,0 @@
-﻿<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE rss [
-  <!ENTITY test "xxx">
-]>
-<rss version="2.0">
-  <channel>
-    <item>
-      <title>&test;</title>
-      <link>&test;</link>
-      <description>
-        &test;
-      </description>
-    </item>
-  </channel>
-</rss>
